+++ stac_fastapi/types/stac_fastapi/types/core.py
@@ -306,6 +306,13 @@ class LandingPageMixin(abc.ABC):
                     "href": urljoin(base_url, "search"),
                     "method": "POST",
                 },
+                {
+                    "rel": Relations.queryables.value,
+                    "type": MimeTypes.jsonschema,
+                    "title": "Queryables",
+                    "href": urljoin(base_url, "queryables"),
+                    "method": "GET",
+                },
             ],
             stac_extensions=extension_schemas,
         )